+++ src/python/pants/backend/python/util_rules/dists.py
@@ -14,15 +14,13 @@ import toml
 from pants.backend.python.subsystems.setup import PythonSetup
 from pants.backend.python.subsystems.setuptools import Setuptools
 from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
-from pants.backend.python.util_rules.pex import (
+from pants.backend.python.util_rules.pex import PexRequest, VenvPex, VenvPexProcess
+from pants.backend.python.util_rules.pex import rules as pex_rules
+from pants.backend.python.util_rules.pex_requirements import (
     Lockfile,
     LockfileContent,
-    PexRequest,
     PexRequirements,
-    VenvPex,
-    VenvPexProcess,
 )
-from pants.backend.python.util_rules.pex import rules as pex_rules
 from pants.base.glob_match_error_behavior import GlobMatchErrorBehavior
 from pants.engine.fs import (
     CreateDigest,
