+++ src/python/pants/backend/python/util_rules/pex.py
@@ -22,7 +22,6 @@ from pants.backend.python.subsystems.repos import PythonRepos
 from pants.backend.python.subsystems.setup import InvalidLockfileBehavior, PythonSetup
 from pants.backend.python.target_types import MainSpecification, PexLayout
 from pants.backend.python.target_types import PexPlatformsField as PythonPlatformsField
-from pants.backend.python.target_types import PythonRequirementsField
 from pants.backend.python.util_rules import pex_cli
 from pants.backend.python.util_rules.interpreter_constraints import InterpreterConstraints
 from pants.backend.python.util_rules.lockfile_metadata import (
@@ -36,6 +35,11 @@ from pants.backend.python.util_rules.pex_environment import (
     PexRuntimeEnvironment,
     PythonExecutable,
 )
+from pants.backend.python.util_rules.pex_requirements import Lockfile, LockfileContent
+from pants.backend.python.util_rules.pex_requirements import (
+    PexRequirements as PexRequirements,  # Explicit re-export.
+)
+from pants.backend.python.util_rules.pex_requirements import ToolCustomLockfile, ToolDefaultLockfile
 from pants.core.util_rules.lockfile_metadata import InvalidLockfileError
 from pants.engine.collection import Collection, DeduplicatedCollection
 from pants.engine.engine_aware import EngineAwareParameter
@@ -57,91 +61,9 @@ from pants.util.docutil import doc_url
 from pants.util.frozendict import FrozenDict
 from pants.util.logging import LogLevel
 from pants.util.meta import frozen_after_init
-from pants.util.ordered_set import FrozenOrderedSet
 from pants.util.strutil import pluralize
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass(frozen=True)
-class Lockfile:
-    file_path: str
-    file_path_description_of_origin: str
-    lockfile_hex_digest: str | None
-    req_strings: FrozenOrderedSet[str] | None
-
-
-@dataclass(frozen=True)
-class LockfileContent:
-    file_content: FileContent
-    lockfile_hex_digest: str | None
-    req_strings: FrozenOrderedSet[str] | None
-
-
-@dataclass(frozen=True)
-class _ToolLockfileMixin:
-    options_scope_name: str
-    uses_source_plugins: bool
-    uses_project_interpreter_constraints: bool
-
-
-@dataclass(frozen=True)
-class ToolDefaultLockfile(LockfileContent, _ToolLockfileMixin):
-    pass
-
-
-@dataclass(frozen=True)
-class ToolCustomLockfile(Lockfile, _ToolLockfileMixin):
-    pass
-
-
-@frozen_after_init
-@dataclass(unsafe_hash=True)
-class PexRequirements:
-    req_strings: FrozenOrderedSet[str]
-    constraints_strings: FrozenOrderedSet[str]
-    # TODO: The constraints.txt resolve for `resolve_all_constraints` will be removed as part of
-    # #12314, but in the meantime, it "acts like" a lockfile, but isn't actually typed as a Lockfile
-    # because the constraints are modified in memory first. This flag marks a `PexRequirements`
-    # resolve as being a request for the entire constraints file.
-    is_all_constraints_resolve: bool
-    repository_pex: Pex | None
-
-    def __init__(
-        self,
-        req_strings: Iterable[str] = (),
-        *,
-        constraints_strings: Iterable[str] = (),
-        is_all_constraints_resolve: bool = False,
-        repository_pex: Pex | None = None,
-    ) -> None:
-        """
-        :param req_strings: The requirement strings to resolve.
-        :param constraints_strings: Constraints strings to apply during the resolve.
-        :param repository_pex: An optional PEX to resolve requirements from via the Pex CLI
-            `--pex-repository` option.
-        """
-        self.req_strings = FrozenOrderedSet(sorted(req_strings))
-        self.constraints_strings = FrozenOrderedSet(sorted(constraints_strings))
-        self.is_all_constraints_resolve = is_all_constraints_resolve
-        self.repository_pex = repository_pex
-
-    @classmethod
-    def create_from_requirement_fields(
-        cls,
-        fields: Iterable[PythonRequirementsField],
-        constraints_strings: Iterable[str],
-        *,
-        additional_requirements: Iterable[str] = (),
-    ) -> PexRequirements:
-        field_requirements = {str(python_req) for field in fields for python_req in field.value}
-        return PexRequirements(
-            {*field_requirements, *additional_requirements},
-            constraints_strings=constraints_strings,
-        )
-
-    def __bool__(self) -> bool:
-        return bool(self.req_strings)
 
 
 class PexPlatforms(DeduplicatedCollection[str]):
